Rename Input Activity into Generate Activity, and integrate Logic methods into Generate Activity
